Refactor client suite in line with server suites
