+++ lib/bulk_data_test_kit.rb
@@ -2,4 +2,4 @@
 require_relative 'inferno_requirements_tools/ext/inferno_core/runnable'
 require_relative 'bulk_data_test_kit/v1.0.1/bulk_data_test_suite'
 require_relative 'bulk_data_test_kit/v2.0.0/bulk_data_test_suite'
-require_relative 'bulk_data_test_kit/v2.0.0/client/bulk_data_client_test_suite'
+require_relative 'bulk_data_test_kit/v2.0.0_client/bulk_data_client_test_suite'
