--- lib/bulk_data_test_kit/v2.0.0/client/tags.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module BulkDataTestKit
-  module BulkDataV200
-    module Client
-      # Tags for use in Bulk Data Client tests and endpoints
-      module Tags
-        METADATA_TAG = 'bulk_data_metadata'
-        PATIENT_KICKOFF_TAG = 'bulk_data_kickoff_patient'
-        GROUP_KICKOFF_TAG = 'bulk_data_kickoff_group'
-        SYSTEM_KICKOFF_TAG = 'bulk_data_kickoff_system'
-        STATUS_TAG = 'bulk_data_status'
-        OUTPUT_TAG = 'bulk_data_output'
-        DELETE_TAG = 'bulk_data_delete'
-      end
-    end
-  end
-end
