+++ lib/bulk_data_test_kit/v2.0.0_client/export_types.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module BulkDataTestKit
+  module BulkDataV200Client
+    # Export types for use in Bulk Data Client tests and endpoints
+    module ExportTypes
+      PATIENT_EXPORT_TYPE = 'Patient'
+      GROUP_EXPORT_TYPE = 'Group'
+      SYSTEM_EXPORT_TYPE = 'System'
+    end
+  end
+end
